\name{pgrid}
\alias{pgrid}
\title{
Lay a grid over a plot
}
\description{
Lays a grid over a plot to know positions
}
\usage{
pgrid(n=10,cex=0.5)
}
\arguments{
  \item{n}{Number of points in each direction}
  \item{cex}{Character expansion}
}
\value{
A grid layed on a plot.
}
\author{
Francois Rousseu
}

