\name{bbox2pol}
\alias{bbox2pol}

\title{
Bounding box 2 SpatialPolygons
}
\description{
Turns a bounding box to a \code{\link{SpatialPolygons}}
}
\usage{
bbox2pol(x,ex=NULL)
}

\arguments{
  \item{x}{A spatial object}
  \item{ex}{A value in the same unit as the spatial object by which to extend the box}

}
\value{
A \code{\link{SpatialPolygons}}
}


